backend/requirements.txt
accelerate
httpx
databases
diffusers>=0.27.0
fastapi
gfpgan
numpy==1.24.4
onnxruntime-gpu==1.15.0
Pillow
#PyMySQL
python-dotenv
python-multipart
rembg
requests
safetensors
torch>=2.0.0
torchvision>=0.17.0
transformers
uvicorn
openai>=1.0.0
git+https://github.com/xinntao/Real-ESRGAN.git
aiomysql

